adding checks for undefined..

+++ frontend/src/components/gmp/PenRequestComplete.vue
@@ -273,12 +273,14 @@ export default {
     notifications() {
       let outcome = null;
       let notification = this.$store.getters['notifications/notification'];
-      notification = JSON.parse(notification);
-      if (notification && notification.penRequestID === this.requestId && notification.sagaStatus === 'COMPLETED' && notification.sagaName === 'PEN_REQUEST_UNLINK_SAGA') {
-        this.loadPenRequest();
-        outcome = 'SAGA_COMPLETED';
-        // eslint-disable-next-line
-        this.unlinkMessage ='Your request to unlink is now completed.';
+      if (notification) {
+        notification = JSON.parse(notification);
+        if (notification && notification.penRequestID === this.requestId && notification.sagaStatus === 'COMPLETED' && notification.sagaName === 'PEN_REQUEST_UNLINK_SAGA') {
+          this.loadPenRequest();
+          outcome = 'SAGA_COMPLETED';
+          // eslint-disable-next-line
+          this.unlinkMessage = 'Your request to unlink is now completed.';
+        }
       }
       return outcome;
     },
